Add files via upload
I noticed that changing the size of the browser window will impact the layout of the content a lot because there are fixed margins but the size of the boxes is undefined. I think we should include a fixed width, at least for the menu, in the CSS so that each menu item fits on one line.